backend/requirements.txt
grpcio
grpcio-tools
pika
flask
flask-cors
redis
